+++ packages/karpotic/src/Block/Block.types.ts
@@ -1,15 +1,7 @@
-import { ComponentProps, FC } from "react";
-import { Color, DimensionValue, SpacingValue } from "../types";
+import { FC } from "react";
+import { ContainerProps } from "../Container";
 
-interface BlockOwnProps {
-  w?: DimensionValue;
-  h?: DimensionValue;
-  p?: SpacingValue;
-  m?: SpacingValue;
-  bgc?: Color;
-}
-
-export type BlockProps = ComponentProps<"div"> & BlockOwnProps;
+export type BlockProps = ContainerProps;
 
 export interface BlockCompoundComponents {
   Center: FC<BlockProps>;
