+++ packages/karpotic/src/Container/Container.tsx
@@ -1,0 +1,27 @@
+import { FC } from "react";
+import {
+  createStyleFromBackgroundColor,
+  createStyleFromDimensions,
+} from "../utils";
+import { ContainerProps } from "./Container.types";
+
+const Container: FC<ContainerProps> = ({ w, h, p, m, bgc, ...props }) => {
+  const { style, children, ...restProps } = props;
+  const dimensionStyles = createStyleFromDimensions(w, h, p, m);
+  const bgcStyle = createStyleFromBackgroundColor(bgc);
+
+  return (
+    <div
+      {...restProps}
+      style={{
+        ...style,
+        ...dimensionStyles,
+        ...bgcStyle,
+      }}
+    >
+      {children}
+    </div>
+  );
+};
+
+export default Container;
