+++ packages/karpotic/src/Container/index.ts
@@ -1,0 +1,2 @@
+export { default as Container } from "./Container";
+export type { ContainerProps } from "./Container.types";
